app/requirements.txt
discord.py
requests
pyyaml
secrets
mysql.connector
